requirements.txt
Babel==2.5.1
Django==1.11
django-phonenumber-field==1.3.0
djangorestframework==3.7.3
phonenumberslite==8.8.8
pytz==2017.3
